AudioPlayback_Status moved to MusicCommon.h
It's used both in ISteamMusic and IClientMusic.
This fixes "unknown type name 'AudioPlayback_Status'" in ISteamMusic.
